Tidy up Quaternion implementation
As with Vector3, the keyword "self" has been removed from the code
wherever it was not strictly needed and did not significantly assist
with readability.

The "@inlinable" attribute has also been added to several simple methods
and properties. These are mostly places where the functionality is
delegated to other methods and properties, and so can be considered as
shorthands.